requirements.txt
python-dotenv
schedule
pymsteams